+++ src/lib.rs
@@ -127,6 +127,8 @@ extern crate netcdf;
 #[cfg(feature = "dataframe")]
 extern crate json;
 
+extern crate order_stat;
+
 pub mod statistics;
 pub mod structure;
 #[macro_use]
@@ -208,11 +210,9 @@ pub use util::api::*;
 //pub use numerical::gauss_legendre::*;
 
 #[allow(unused_imports)]
-#[cfg(feature = "special")]
 pub use statistics::dist::*;
 
 #[allow(unused_imports)]
-#[cfg(feature = "special")]
 pub use special::function::*;
 
 #[allow(unused_imports)]